Add support for chained fit files
From the SDK:
> The FIT protocol allows for multiple FIT files to be chained together
> in a single FIT file. Each FIT file in the chain must be a properly
> formatted FIT file (header, data records, CRC).

The implementation is to note the total filesize when first loading the
file, then check if there's more data after we finish processing the
number of bytes the initial header specifies. If there is more data, try
to parse it as a fit header, then continue as normal.