Fix Python 2 issue with int/long types
In Python 2, integer objects are automatically promoted to objects of
type `long` if they're too large for a normal `int`. In Python 3, all
integers are `int` objects, no matter how big.

This has the effect of messing up the `isinstance` checks that only
check for `int` and not `long`.

This commit fixes this issue by defining a different list of types to
check for depending on the Python version.

Fixes #26