test: clarify test_fileish_types
There was a naming problem on case-insensitive filesystems, which is why
the tests refered to both 'Settings.FIT' and 'Settings.FIT', with the
latter not inside git, and tests failing on all case-sensitive
filesystems. Fix this by duplicating 'Settings.fit' to 'nametest.FIT',
and using it inside `test_fileish_types(self)`. Also improve test
documentation a bit, while adding a link to the underlying issue.